+++ ESFramework/UIKit/UIToolbar+ESAdditions.h
@@ -13,7 +13,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface UIToolbar (ESAdditions)
 
 - (nullable UIBarButtonItem *)itemWithTag:(NSInteger)tag;
-- (void)replaceItemWithTag:(NSInteger)tag toItem:(UIBarButtonItem *)newItem;
+- (void)replaceItemWithTag:(NSInteger)tag toItem:(UIBarButtonItem *)newItem animated:(BOOL)animated;
 
 @end